split Lesson.tsx components into separate files as suggested
